Async loader fixes
* fix `fn load_from_disk` -> `async fn load_from_disk`

Old code was incompatible with async and because this code here...
```rust
    for (path, handle) in handles.drain(..) {
        let bytes = handle
            .join()
```
... blocks the calling thread until all threads are joined (because join is a blocking call), which in the case of a single-threaded runtime would have the effect of preventing any other tasks/`async fn` from executing. In a multi-threaded runtime, it would just block all tasks on that particular thread, which could be a lot. This might exhibit as lag in other tasks while loading is in progress.

New code for non-wasm uses tokio tasks (green threads) and lets the runtime decide how to schedule file io by `await`ing instead of `join`ing. This allows the runtime to schedule other tasks for execution while awaiting the tasks. Note that although the tokio runtime will likely decide to load from disk from a single worker thread in series. The difference here is that the runtime knows it's going on and will do all that in a dedicated thread, which won't block the calling thread.

* new non-wasm download code

The old code used async and await, and the code was async, but the downloads would still happen sequentially because when you await in a loop, it stops iteration *within that function*. In the following code, each `send()` would therefore be made in sequence. Other async functions could be interleaved and run, but based on the name `handles` I think the intent here was to await all paths concurrently.
```rust
        for path in paths {
            let url = reqwest::Url::parse(path.to_str().unwrap())
                .map_err(|_| Error::FailedParsingUrl(path.to_str().unwrap().to_string()))?;
            handles.push((path, client.get(url).send().await));
```

The new code uses tokio tasks, which are all awaited concurrently -- the only limit being 8 connections per host, to avoid hammering servers.

* simplify `load_single`

`load_single` now just calls `load_async_single` from a new executor on the current thread. Async and sync code is shared this way and performance will likely benefit provided it's not called in a tight loop. Tasks within the future may still be awaited concurrently provided they don't block.

* ignore `.vscode`
* enable parallel network and disk loading

Old behavior was to await network assets, then await disk assets.

New behavior is to await both disk and network IO concurrently.

https://rust-lang.github.io/async-book/06_multiple_futures/02_join.html

* simplify `load` code
* document `load_async_single`
* add `RawAssets::with_capacity` - to allow pre-allocating space when the len is known ahead of time